create Teletubby component
