+++ src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Layout from "pages/Layout";
 import Home from "pages/Home";
 import NoPage from "pages/NoPage";
-import Teletubbies from "components/Teletubbies";
+import Teletubbies from "pages/Teletubbies";
 import "./App.css";
 function App() {
   return (
